test: new test organization based on the interfaces
