--- test/fs_suite_linux.go
@@ -1,10 +1,0 @@
-// +build !windows
-
-package test
-
-import "os"
-
-var (
-	customMode            os.FileMode = 0755
-	expectedSymlinkTarget             = "/dir/file"
-)
